# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Evercode Lab
  version: 1.0.0

servers:
  - url: http://localhost:3001

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

  schemas:
    Currency:
      type: object
      required:
        - name
        - ticker
      properties:
        name:
          type: string
          example: Bitcoin
        ticker:
          type: string
          example: BTC

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Forbidden
        context:
          type: object
          additionalProperties: true

paths:
  /status:
    get:
      summary: Check server status
      responses:
        '200':
          description: Server is running
          content:
            text/plain:
              schema:
                type: string
                example: ok

  /currencies:
    get:
      summary: Get all currencies
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of currencies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Currency'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create currency
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Currency'
      responses:
        '201':
          description: Currency created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Currency already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /currencies/{ticker}:
    get:
      summary: Get currency by ticker
      security:
        - BearerAuth: []
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
            example: BTC
      responses:
        '200':
          description: Currency found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Currency not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update currency by ticker
      security:
        - BearerAuth: []
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
          example: BTC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Currency'
      responses:
        '200':
          description: Currency updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Currency not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete currency by ticker
      security:
        - BearerAuth: []
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
          example: BTC
      responses:
        '204':
          description: Currency deleted
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Currency not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'